refactor: regroup the permission retrieval functions into one that handles all cases
