+++ connectors/src/connectors/zendesk/lib/brand_permissions.ts
@@ -1,17 +1,10 @@
-import type {
-  ConnectorPermission,
-  ContentNode,
-  ContentNodesViewType,
-  ModelId,
-} from "@dust-tt/types";
+import type { ModelId } from "@dust-tt/types";
 
 import { allowSyncZendeskHelpCenter } from "@connectors/connectors/zendesk/lib/help_center_permissions";
-import { getBrandInternalId } from "@connectors/connectors/zendesk/lib/id_conversions";
 import { allowSyncZendeskTickets } from "@connectors/connectors/zendesk/lib/ticket_permissions";
 import { getZendeskAccessToken } from "@connectors/connectors/zendesk/lib/zendesk_access_token";
 import { createZendeskClient } from "@connectors/connectors/zendesk/lib/zendesk_api";
 import logger from "@connectors/logger/logger";
-import { ConnectorResource } from "@connectors/resources/connector_resource";
 import { ZendeskBrandResource } from "@connectors/resources/zendesk_resources";
 
 /**
@@ -105,57 +98,4 @@ export async function revokeSyncZendeskBrand({
 
   await brand.revokePermissions();
   return brand;
-}
-
-export async function retrieveZendeskBrandPermissions({
-  connectorId,
-  parentInternalId,
-  filterPermission,
-}: {
-  connectorId: ModelId;
-  parentInternalId: string | null;
-  filterPermission: ConnectorPermission | null;
-  viewType: ContentNodesViewType;
-}): Promise<ContentNode[]> {
-  const connector = await ConnectorResource.fetchById(connectorId);
-  if (!connector) {
-    logger.error({ connectorId }, "[Zendesk] Connector not found.");
-    throw new Error("Connector not found");
-  }
-
-  const isReadPermissionsOnly = filterPermission === "read";
-  const isRootLevel = !parentInternalId;
-  let nodes: ContentNode[] = [];
-
-  // At the root level, we show one node for each brand.
-  if (isRootLevel) {
-    if (isReadPermissionsOnly) {
-      const brandsInDatabase =
-        await ZendeskBrandResource.fetchBrandsWithHelpCenter({ connectorId });
-      nodes = brandsInDatabase.map((brand) =>
-        brand.toContentNode({ connectorId })
-      );
-    } else {
-      const token = await getZendeskAccessToken(connector.connectionId);
-      const zendeskApiClient = createZendeskClient({ token });
-
-      const { result: brands } = await zendeskApiClient.brand.list();
-      nodes = brands
-        .filter((brand) => brand.has_help_center)
-        .map((brand) => ({
-          provider: connector.type,
-          internalId: getBrandInternalId(connectorId, brand.id),
-          parentInternalId: null,
-          type: "folder",
-          title: brand.name || "Brand",
-          sourceUrl: brand.brand_url,
-          expandable: true,
-          permission: "none",
-          dustDocumentId: null,
-          lastUpdatedAt: null,
-        }));
-    }
-  }
-  nodes.sort((a, b) => a.title.localeCompare(b.title));
-  return nodes;
 }
